+++ packages/bitbucket/panel/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/panel
 
+## 0.0.9
+- Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
+  - @atlaskit/docs@5.2.2
+  - @atlaskit/button@10.0.1
+  - @atlaskit/icon@15.0.0
+
 ## 0.0.8
 - Updated dependencies [6998f11](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6998f11):
   - @atlaskit/docs@5.2.1