+++ packages/core/banner/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/banner
 
+## 7.0.10
+- Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
+  - @atlaskit/docs@5.2.2
+  - @atlaskit/button@10.0.1
+  - @atlaskit/icon@15.0.0
+
 ## 7.0.9
 - Updated dependencies [6998f11](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6998f11):
   - @atlaskit/docs@5.2.1