+++ packages/core/button/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/button
 
+## 10.0.2
+- [patch] [8f89287](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8f89287):
+
+  - Add tslib to dependencies to stop load breaking when it's not there
+
+## 10.0.1
+- Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
+  - @atlaskit/docs@5.2.2
+  - @atlaskit/checkbox@5.0.7
+  - @atlaskit/icon@15.0.0
+
 ## 10.0.0
 - [major] [6998f11](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6998f11):