+++ packages/core/modal-dialog/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/modal-dialog
 
+## 7.0.13
+- Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
+  - @atlaskit/docs@5.2.2
+  - @atlaskit/avatar@14.1.5
+  - @atlaskit/button@10.0.1
+  - @atlaskit/checkbox@5.0.7
+  - @atlaskit/field-radio-group@4.0.12
+  - @atlaskit/inline-dialog@9.0.11
+  - @atlaskit/portal@0.0.16
+  - @atlaskit/icon@15.0.0
+
 ## 7.0.12
 - Updated dependencies [6998f11](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6998f11):
   - @atlaskit/docs@5.2.1