+++ packages/core/navigation/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @atlaskit/navigation
 
+## 33.3.6
+- Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
+  - @atlaskit/docs@5.2.2
+  - @atlaskit/avatar@14.1.5
+  - @atlaskit/button@10.0.1
+  - @atlaskit/checkbox@5.0.7
+  - @atlaskit/dropdown-menu@6.1.23
+  - @atlaskit/item@8.0.12
+  - @atlaskit/section-message@1.0.12
+  - @atlaskit/single-select@6.0.9
+  - @atlaskit/toggle@5.0.12
+  - @atlaskit/tooltip@12.1.11
+  - @atlaskit/quick-search@5.1.2
+  - @atlaskit/icon@15.0.0
+
 ## 33.3.5
 - Updated dependencies [6998f11](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6998f11):
   - @atlaskit/docs@5.2.1