+++ packages/core/tooltip/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/tooltip
 
+## 12.1.11
+- Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
+  - @atlaskit/docs@5.2.2
+  - @atlaskit/button@10.0.1
+  - @atlaskit/portal@0.0.16
+  - @atlaskit/icon@15.0.0
+
 ## 12.1.10
 - Updated dependencies [6998f11](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6998f11):
   - @atlaskit/docs@5.2.1