+++ packages/editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-mobile-bridge
 
+## 6.3.2
+- [patch] [3061b52](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3061b52):
+
+  - AK-5723 - adjust files in package.json to ensure correct publishing of dist/package.json
+
 ## 6.3.1
 - [patch] [345b45c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/345b45c):