+++ packages/editor/editor-test-helpers/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @atlaskit/editor-test-helpers
 
+## 6.2.22
+- Updated dependencies [df32968](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/df32968):
+  - @atlaskit/editor-core@89.0.7
+  - @atlaskit/smart-card@9.0.0
+
+## 6.2.21
+- Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
+  - @atlaskit/smart-card@8.8.5
+  - @atlaskit/editor-common@22.0.2
+  - @atlaskit/editor-core@89.0.6
+  - @atlaskit/media-test-helpers@18.3.1
+  - @atlaskit/icon@15.0.0
+
 ## 6.2.20
 - [patch] [a2cae0c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a2cae0c):