+++ packages/elements/reactions/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @atlaskit/reactions
 
+## 15.0.7
+- [patch] [3061b52](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3061b52):
+
+  - AK-5723 - adjust files in package.json to ensure correct publishing of dist/package.json
+
+## 15.0.6
+- Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
+  - @atlaskit/docs@5.2.2
+  - @atlaskit/button@10.0.1
+  - @atlaskit/tooltip@12.1.11
+  - @atlaskit/emoji@47.0.2
+  - @atlaskit/icon@15.0.0
+
 ## 15.0.5
 - Updated dependencies [6998f11](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6998f11):
   - @atlaskit/docs@5.2.1