+++ packages/jira/color-picker/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/color-picker
 
+## 0.1.2
+- [patch] [3061b52](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3061b52):
+
+  - AK-5723 - adjust files in package.json to ensure correct publishing of dist/package.json
+
+## 0.1.1
+- Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
+  - @atlaskit/docs@5.2.2
+  - @atlaskit/section-message@1.0.12
+  - @atlaskit/select@6.1.9
+  - @atlaskit/icon@15.0.0
+
 ## 0.1.0
 - [minor] [b5df1db](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b5df1db):